requirements.txt
faster-whisper
fastapi
uvicorn
python-dotenv
openai
librosa
soundfile
torch
torchaudio
ffmpeg-python
googlemaps
